requirements.txt
discord.py[voice]==2.3.2
yt-dlp==2025.11.3.233024.dev0
PyNaCl==1.5.0
python-dotenv==1.0.0
urllib3==1.26.18
browser_cookie3==0.19.1